requirements.txt
streamlit==1.11.1
pandas==1.2.3
base58==2.1.0
streamlit-option-menu==0.3.2
pillow==8.2.0
scipy==1.8.1
seaborn==0.11.2
matplotlib==3.5.2
openpyxl==3.0.10